+++ core/src/main/scala/com/krux/hyperion/DataPipelineDef.scala
@@ -4,10 +4,17 @@ import scala.language.implicitConversions
 
 import com.amazonaws.services.datapipeline.model.{ ParameterObject => AwsParameterObject,
   PipelineObject => AwsPipelineObject }
-import org.json4s.JsonDSL._
-import org.json4s.{ JArray, JValue }
 
-import com.krux.hyperion.aws.{ AdpJsonSerializer, AdpParameterSerializer, AdpPipelineSerializer }
+import com.amazonaws.services.datapipeline.model.{ParameterObject => AwsParameterObject, PipelineObject => AwsPipelineObject}
+import com.krux.hyperion.activity.MainClass
+import com.krux.hyperion.aws.{AdpJsonSerializer, AdpParameterSerializer, AdpPipelineSerializer}
+import com.krux.hyperion.common._
+import com.krux.hyperion.expression.{Parameter, ParameterValues}
+import com.krux.hyperion.workflow.WorkflowExpressionImplicits
+import org.json4s.JsonDSL._
+import org.json4s.{JArray, JValue}
+
+import com.krux.hyperion.aws.{AdpJsonSerializer, AdpParameterSerializer, AdpPipelineSerializer}
 import com.krux.hyperion.workflow.WorkflowExpression
 
 
@@ -29,15 +36,13 @@ object DataPipelineDef {
 
   implicit def dataPipelineDef2Json(pd: DataPipelineDef): JValue =
     ("objects" -> JArray(
-      AdpJsonSerializer(pd.defaultObject.serialize) ::
-      AdpJsonSerializer(pd.schedule.serialize) ::
-      pd.workflow.toPipelineObjects.map(_.serialize).toList.sortBy(_.id).map(o => AdpJsonSerializer(o)))) ~
+      (pd.defaultObject +: pd.schedule +: pd.workflow.toPipelineObjects.toList)
+        .map(_.serialize).sortBy(_.id).map(o => AdpJsonSerializer(o)))) ~
     ("parameters" -> JArray(
       pd.parameters.flatMap(_.serialize).map(o => AdpJsonSerializer(o)).toList))
 
   implicit def dataPipelineDef2Aws(pd: DataPipelineDef): Seq[AwsPipelineObject] =
-    AdpPipelineSerializer(pd.defaultObject.serialize) ::
-    AdpPipelineSerializer(pd.schedule.serialize) ::
-    pd.workflow.toPipelineObjects.map(o => AdpPipelineSerializer(o.serialize)).toList
+    (pd.defaultObject +: pd.schedule +: pd.workflow.toPipelineObjects.toList)
+      .map(o => AdpPipelineSerializer(o.serialize))
 
 }