+++ core/src/main/scala/com/krux/hyperion/aws/AdpDataPipelineObject.scala
@@ -14,7 +14,7 @@ abstract class AdpDataPipelineAbstractObject extends AdpObject {
 trait AdpDataPipelineDefaultObject extends AdpDataPipelineAbstractObject {
   val id: String = "Default"
   val name: Option[String] = Option("Default")
-  def fields: Map[String, Either[String, AdpRef[AdpDataPipelineObject]]]
+  def fields: Map[String, Either[String, AdpRef[AdpDataPipelineAbstractObject]]]
 }
 
 /**