+++ core/src/main/scala/com/krux/hyperion/cli/AwsAction.scala
@@ -1,7 +1,7 @@
 package com.krux.hyperion.cli
 
+import com.krux.hyperion.client.{AwsClientForDef, AwsClient}
 import com.krux.hyperion.DataPipelineDefGroup
-import com.krux.hyperion.client.{ AwsClientForDef, AwsClient }
 
 private[hyperion] trait AwsAction extends Action {