+++ core/src/main/scala/com/krux/hyperion/resource/SparkCluster.scala
@@ -13,7 +13,7 @@ case class SparkCluster private (
   baseFields: BaseFields,
   resourceFields: ResourceFields,
   emrClusterFields: EmrClusterFields,
-  sparkVersion: HString
+  sparkVersion: Option[HString]
 ) extends EmrCluster {
 
   type Self = SparkCluster
@@ -24,11 +24,17 @@ case class SparkCluster private (
   def updateResourceFields(fields: ResourceFields) = copy(resourceFields = fields)
   def updateEmrClusterFields(fields: EmrClusterFields) = copy(emrClusterFields = fields)
 
-  def withSparkVersion(sparkVersion: HString) = copy(sparkVersion = sparkVersion)
+  def withSparkVersion(sparkVersion: HString) = copy(sparkVersion = Option(sparkVersion))
 
-  override def standardBootstrapAction = 
-    (s"s3://support.elasticmapreduce/spark/install-spark,-v,${sparkVersion},-x": HString) +:
+  override def applications = if (releaseLabel.nonEmpty)
+    ("Spark": HString) +: super.applications
+  else
+    super.applications
+
+  override def standardBootstrapAction = if (releaseLabel.nonEmpty)
     super.standardBootstrapAction
+  else
+    sparkVersion.map(version => s"s3://support.elasticmapreduce/spark/install-spark,-v,${version},-x": HString).toSeq ++ super.standardBootstrapAction
 
 }
 
@@ -38,7 +44,7 @@ object SparkCluster {
     baseFields = BaseFields(PipelineObjectId(SparkCluster.getClass)),
     resourceFields = EmrCluster.defaultResourceFields(hc),
     emrClusterFields = EmrCluster.defaultEmrClusterFields(hc),
-    sparkVersion = hc.emrSparkVersion.get
+    sparkVersion = hc.emrSparkVersion
   )
 
 }