+++ core/src/main/scala/com/krux/hyperion/workflow/WorkflowExpressionImplicits.scala
@@ -16,4 +16,8 @@ trait WorkflowExpressionImplicits {
   implicit def activity2WorkflowExpression(activity: PipelineActivity[_ <: ResourceObject]): WorkflowExpression =
     WorkflowActivityExpression(activity)
 
+  implicit class activityWorkflowExpressionOps(activity: PipelineActivity[_ <: ResourceObject]) {
+    def toWorkflowExpression: WorkflowExpression = activity
+  }
+
 }